Feature/memo

+++ src/main/java/com/fatec/sasbackend/product/ProductServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
 import java.util.Objects;
 
 
@@ -25,7 +26,6 @@ public class ProductServiceImpl implements ProductService {
     public Page<ProductDTO> findAllProducts(Pageable pageable) {
         return productRepository.findAll(pageable)
                 .map(product -> productConverter.fromEntityToDto(new ProductDTO(), product));
-
     }
 
     @Override
@@ -79,6 +79,23 @@ public class ProductServiceImpl implements ProductService {
         Product entity = productRepository.findById(Long.parseLong(id))
                 .orElseThrow(() -> new NotFoundException(PRODUCT_NOT_FOUND));
 
+        if(entity.getAvailableQuantity() > 0) {
+            throw new BadRequestException("Produtos com quantidades disponiveis não podem ser excluidos do estoque!");
+        }
+
         productRepository.deleteById(Long.parseLong(id));
+    }
+
+
+    @Override
+    public List<SimpleProductDTO> findProductsForBeneficiary() {
+        return productRepository.findAll()
+                .stream()
+                .map(product -> SimpleProductDTO.builder()
+                        .id(product.getId())
+                        .name(product.getName())
+                        .build()
+                )
+                .toList();
     }
 }
